rename CSRSpec.hs into CSRGetSet.hs and wrap it in CSRSpec.hs
to do the fromIntegral conversions once and for all
