+++ src/Spec/ExecuteCSR.hs
@@ -23,36 +23,34 @@ execute (Csrrw rd rs1 csr12) = do
   x <- getRegister rs1
   y <- getCSR (lookupCSR csr12)
   setCSR (lookupCSR csr12) x
-  when (rd /= 0) (do
-    setRegister rd ((fromIntegral:: MachineInt -> t) y))
+  when (rd /= 0) (setRegister rd y)
 execute (Csrrs rd rs1 csr12) = do
   checkPermissions (rs1 /= 0) csr12
   mask <- getRegister rs1
   val <- getCSR (lookupCSR csr12)
-  when (rs1 /= 0) (setCSR (lookupCSR csr12) (((fromIntegral:: MachineInt -> t) val) .|. mask))
-  setRegister rd ((fromIntegral:: MachineInt -> t) val)
+  when (rs1 /= 0) (setCSR (lookupCSR csr12) (val .|. mask))
+  setRegister rd val
 execute (Csrrc rd rs1 csr12) = do
   checkPermissions (rs1 /= 0) csr12
   mask <- getRegister rs1
   val <- getCSR (lookupCSR csr12)
-  when (rs1 /= 0) (setCSR (lookupCSR csr12) (((fromIntegral:: MachineInt -> t) val) .&. (complement mask)))
-  setRegister rd ((fromIntegral:: MachineInt -> t) val)
+  when (rs1 /= 0) (setCSR (lookupCSR csr12) (val .&. (complement mask)))
+  setRegister rd val
 execute (Csrrwi rd zimm csr12) = do
   checkPermissions True csr12
   val <- getCSR (lookupCSR csr12)
   setCSR (lookupCSR csr12) ((fromIntegral:: MachineInt -> t) zimm)
-  when (rd /= 0) (do
-    setRegister rd ((fromIntegral:: MachineInt -> t) val))
+  when (rd /= 0) (setRegister rd val)
 execute (Csrrsi rd zimm csr12) = do
   checkPermissions (zimm /= 0) csr12
   val <- getCSR (lookupCSR csr12)
-  when (zimm /= 0) (setCSR (lookupCSR csr12) (((fromIntegral:: MachineInt -> t) val) .|. ((fromIntegral:: MachineInt -> t) zimm)))
-  setRegister rd ((fromIntegral:: MachineInt -> t) val)
+  when (zimm /= 0) (setCSR (lookupCSR csr12) (val .|. ((fromIntegral:: MachineInt -> t) zimm)))
+  setRegister rd val
 execute (Csrrci rd zimm csr12) = do
   checkPermissions (zimm /= 0) csr12
   val <- getCSR (lookupCSR csr12)
-  when (zimm /= 0) (setCSR (lookupCSR csr12) (((fromIntegral:: MachineInt -> t) val) .&. (complement ((fromIntegral:: MachineInt -> t) zimm))))
-  setRegister rd ((fromIntegral:: MachineInt -> t) val)
+  when (zimm /= 0) (setCSR (lookupCSR csr12) (val .&. (complement ((fromIntegral:: MachineInt -> t) zimm))))
+  setRegister rd val
 execute Ecall = do
   mode <- getPrivMode
   case mode of
